creating wrapper.js as seperate file
